requirements.txt
playwright==1.41.1
matplotlib==3.8.2
pandas==2.1.4
Flask==3.0.0
openpyxl==3.1.2
mysql-connector-python

